Revert "Divided classes into packages. Added classes Input and Output."
This reverts commit 797316250045ef30b8ed9c68a8dd5c0ce2606fbf.
