+++ src/main/java/Duke.java
@@ -1,22 +1,30 @@
-package duke;
-
-import duke.task.Deadline;
-import duke.task.Event;
-import duke.task.Task;
-import duke.task.Todo;
-import duke.command.Input;
-import duke.command.Output;
-
 import java.util.Arrays;
 import java.util.Scanner;
 
 public class Duke {
     public static final int MAX_NUMBER_OF_TASKS = 100;
 
+    public static void printLines() {
+        String lines = "__________________________________________________";
+        System.out.println(lines);
+    }
+
     public static String truncateInput(String[] s) {
         String[] copy = Arrays.copyOfRange(s, 1, s.length);
         String truncatedInput = String.join(" ", copy);
         return truncatedInput;
+    }
+
+    public static void printGreetings(String greeting) {
+        if (greeting.equals("hello")) {
+            // hello greeting
+            printLines();
+            System.out.println("Hello! I'm Duke" + System.lineSeparator() + "What can I do for you?");
+            printLines();
+        }   else {
+            // bye greeting
+            System.out.println("Bye. Hope to see you again soon!");
+        }
     }
 
     public static void markResponse(int position, int count, Task[] tasks) {
@@ -70,7 +78,7 @@ public class Duke {
 
     public static void main(String[] args) {
         // prints hello greeting
-        Output.printGreetings("hello");
+        printGreetings("hello");
 
         String inputs;
         int count = 0; // how many items in array, 0 at start.
@@ -86,10 +94,10 @@ public class Duke {
             String truncatedInput = truncateInput(words);
             // truncatedInput is the input without the command keywords e.g. todo, list, mark...
 
-            Output.printLines();
+            printLines();
             switch (keyword) {
             case "bye":
-                Output.printGreetings("bye");
+                printGreetings("bye");
                 systemState = false;
                 break;
             case "mark":
@@ -118,7 +126,7 @@ public class Duke {
                 System.out.println("ERROR!");
                 break;
             }   // putting this switch statement into another function causes some variables to not be incremented
-            Output.printLines();
+            printLines();
         }
     }
 }
