+++ src/main/java/Event.java
@@ -1,5 +1,3 @@
-package duke.task;
-
 public class Event extends Task {
 
     protected String at;
