+++ src/main/java/Task.java
@@ -1,5 +1,3 @@
-package duke.task;
-
 public class Task {
     protected String description;
     protected boolean isDone;
