--- src/main/java/duke/command/Input.java
@@ -1,4 +1,0 @@
-package duke.command;
-
-public class Input {
-}
